app_rec/requirements.txt
opencv-python>=4.5.0
depthai>=2.30.0
gpiozero>=2.0
RPi.GPIO>=0.7.0
numpy>=1.20.0 